Implement ES6 Promises and Fetch API in place of jQuery's
jQuery 3 actually already implements the ES6 specification for
promises, but the problem is that ES6 promises, by design, swallow
exceptions and jQuery follows the spec a little too exactly.

So on one hand, it's nice you can asynchronously chain catches
and `then` statements to handle exceptions:

```
promise.resolve().then(() => { throw Err("Rejected!"); })
  .then(() => 1, (err) => 2)
  .then((val) => {
    console.log(val); // => 2
    throw new Error("DEF");
  })
  .catch((err) => 3)
  .then((val) => console.log(val) // => 3)

```

But if you have an unintentional exception, the error gets
swallowed (because it's possible that an error handler might later
be added to the chain).

```
// This doesn't throw any exceptions
promise.resolve().then(() => nonExistentFn());
```

This is bad for debugging. Also it doesn't get reported to Sentry
so if it makes it to prod, we wouldn't even know about it.

To get around this, we're using an ES6 shim Promise class which has
a hook that lets us emit an `unhandledrejection` event if there's an
error). I've also hooked up stuff in `log.ts` to report unhandled
exceptions to Raven. We're also using a shim `fetch` library because
we can use it with shim Promises.

Also, because some browsers implement Promises without an
`unhandledrejection` event -- we also override the native
implementation of Promises solely so we can get proper error
reporting.

In addition, rewrites a big chunk of `json-http` to use new Promises
and adds some new hooks and handlers to give the app more options
for reacting to API call state. Adds tests for `json-http` finally.