Groups data handling
Data is stored in the form of `"FETCHING"|"FETCH_ERROR"|<data>`. For
our purposes, we ignore the previously used data states for saving data
(the global data status handler can show stuff like this -- but we usually
don't need to know whether items are being saved on a case by case basis).

Note that because certain fields of groups are optional (members, tags),
we store these separately in store so we can easily check what's missing
or available.