Simplify analytics
This just calls analytics.page using the route hash a "name" for the
page (so "groups.html#!/event-list/:groupId" instead of "Event List")
and just passes the params (e.g. { groupId: "someId" }) along with
it.

This has the advantage of us no longer needing to remember to hook
up basic analytics.

Note that this means if we change routing in any way, this will alter
how things get tracked in analytics. However, Mixpanel has some okay
tools for aliasing events and whatnot, so it really should just be
something that gets handled there rather than in the app itself.