Lutra-redux serves original Lutra files if available.
Groups lives in lutra-redux, but if the lutra repo is in a sibling
directory to lutra-redux, redux's `make watch` will copy files
from lutra over (useful if you need the settings page for instance).

Need to do `make` inside lutra/esper.com first. Then `make watch`
in lutra-redux. Long term goal is to move everything to lutra-redux,
but this'll do for now.