Make path param types singletons.
Using an instance of a type to tell TS what type to use
isn't very reliable. We just use singletons to designate
how we should clean up various types instead now.

Because TS2.1 ships with Partials and mapped interfaces,
we can infer the shape of what params are optional or
required for a Path from its constructor arguments
now.