Trim period being fetched
Suppose user requests to fetch between days m and n. Old behavior
as that if cached data for an day in between m and n is invalid,
we would fetch the entire period from m to n. New behavior is to
fetch only the smallest continguous period possible that would
encompass all the days.

So if days m + x and n - y are invalid (and n - y > m + x), we only
fetch the period between [m + x, n - y].