Scroll confirms labels
When scrolling past an event, starts a timer that confirms labels
on event after 3s. If the user unmounts the event (e.g. if she
goes to another time period), the timer is stopped. If the event
has been labeled by the time the timeout finishes, nothing happens.

Unconfirmed labels get a small blue line next to them. This line
stays even after auto-confirming (because we want user to have some
visual reference to see what they should be double-checking). It
disappears if user clicks on event title or changes the event's
labels.