Fixes fetch promise-related bug.
The bug occurs when you fetch data for a longer period and then
(before the fetch finishes) navgiate to fetch data for a shorter
period within the longer one. The chart for this shorter period
may never render and stay in a "Calculating" state forever.

Because there is already an API call to fetch the longer period
in progress, the fetch function does not make any API call for
the shorter period. As such, the promise it returns resolves more
or less instantly. This is a problem because we rely on that promise
to tell our worker to re-render the chart. It will attempt to do
so before all the event data is ready and terminate without doing
anything.

Solution is to have the fetch function for a shorter period return
a promise for the longer period's fetch if applicable.