Implicit confirmation of timebomb status
Suppose I have a list of 6 events during setup. I set timebomb equal to
on for event 3. We implicitly assume that the user intended timebombs
to be off for events 1 and 2 (assuming they're not already on for whatever
reason).

This really only matters if the user later defaults timebombs to on.
Default on implies that we want timebombs on for events with no
explicit timebomb state. But this could surprise the user because
they consciously(?) chose to *not* turn timebomb on. And if we send
an e-mail right away, this is not undoable by the user.