Fix issue for recurring event IDs
The front end maintains a map from each the "recurring_event_id"
to a list of per-event IDs. It uses this to update all recurring
event instances in its store when a user tries to update all
instances of a recurring event. However, this map was only created
when user fetched event via a query (e.g. all events from time A
to B) and not when user requested an event by a specific ID.

PR now tracks recurring events requested by ID. Also adds in an
extra bit of robustness so if the map is missing, it doesn't crash.